refactor 	- winningNumber -> Lotto 도메인으로 수정  	- winningTest 분리
