+++ src/main/java/step2/model/Winning.java
@@ -4,22 +4,16 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Winning {
-    private final WinningNumber winningNumber;
+    private final Lotto winningNumber;
     private final Bonus bonus;
 
-    public Winning(WinningNumber winningNumber, Bonus bonus) {
+    public Winning(Lotto winningNumber, Bonus bonus) {
         this.winningNumber = winningNumber;
         this.bonus = bonus;
     }
 
     public List<Integer> getWinningNumber() {
-        List<Integer> results = new ArrayList<>();
-
-        for (LottoNumber number : winningNumber.getWinningNumbers()) {
-            results.add(number.getLottoNumber());
-        }
-
-        return results;
+        return winningNumber.getLottoNumbers();
     }
 
     public int getBonus() {
